Attempted code split, but Dotty compiler crash
