+++ src/main/scala/optics/internal/focus/codegen/AllGenerators.scala
@@ -1,0 +1,5 @@
+package optics.internal.focus.codegen
+
+trait AllGenerators 
+  extends GeneratorLoopModule 
+  with FieldSelectionGeneratorModule
