+++ src/main/scala/optics/internal/focus/parse/AllParsers.scala
@@ -1,0 +1,4 @@
+package optics.internal.focus.parse
+
+trait AllParsers extends ParserLoopModule 
+  with FieldSelectionParserModule
